Added the kafka-to-geode pipeline
The initial code features the geode-to-kafka pipeline. The reverse path
is added to make things complete.

The KafkaJFairyProducerDriver class is responsible to generate fake data
and feed to kafka, KafkaConsumerGeodeLoader will be a daemon that will
sync all incoming kafka messages to geode. No geode server-side code is
necessary.

In addition, a description of the commands to run is created as
runit.txt.